+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.4
 * Changed Error Board hints to be more informative.
+* Changed displaying incorrect command option to be more strict.
 # 0.1.3 (2024-01-31)
 * Added logging monitor resolution to `runtime.log`.
 * Added highlighting of in-game exp tooltip region for debug mode.